storage: conslidate index state with compat helpers
